+++ Source/EventFlow.MsSql.Tests/IntegrationTests/EventStores/MsSqlEventStoreTests.cs
@@ -22,21 +22,17 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
 
-using System.Threading;
-using System.Threading.Tasks;
 using EventFlow.Configuration;
 using EventFlow.EventStores.MsSql;
 using EventFlow.Extensions;
 using EventFlow.MsSql.Extensions;
-using EventFlow.MsSql.Tests.IntegrationTests.QueryHandlers;
-using EventFlow.MsSql.Tests.ReadModels;
-using EventFlow.ReadStores.MsSql.Extensions;
 using EventFlow.TestHelpers.Suites;
 using Helpz.MsSql;
+using NUnit.Framework;
 
-namespace EventFlow.MsSql.Tests.IntegrationTests
+namespace EventFlow.MsSql.Tests.IntegrationTests.EventStores
 {
-    public class MssqlReadModelStoreTests : TestSuiteForReadModelStore
+    public class MsSqlEventStoreTests : TestSuiteForEventStore
     {
         private IMsSqlDatabase _testDatabase;
 
@@ -46,8 +42,7 @@ namespace EventFlow.MsSql.Tests.IntegrationTests
 
             var resolver = eventFlowOptions
                 .ConfigureMsSql(MsSqlConfiguration.New.SetConnectionString(_testDatabase.ConnectionString.Value))
-                .UseMssqlReadModel<MsSqlThingyReadModel>()
-                .AddQueryHandlers(typeof(MsSqlThingyGetQueryHandler))
+                .UseEventStore<MsSqlEventPersistence>()
                 .CreateResolver();
 
             var databaseMigrator = resolver.Resolve<IMsSqlDatabaseMigrator>();
@@ -57,14 +52,10 @@ namespace EventFlow.MsSql.Tests.IntegrationTests
             return resolver;
         }
 
-        protected override Task PurgeTestAggregateReadModelAsync()
+        [TearDown]
+        public void TearDown()
         {
-            return ReadModelPopulator.PurgeAsync<MsSqlThingyReadModel>(CancellationToken.None);
-        }
-
-        protected override Task PopulateTestAggregateReadModelAsync()
-        {
-            return ReadModelPopulator.PopulateAsync<MsSqlThingyReadModel>(CancellationToken.None);
+            _testDatabase.Dispose();
         }
     }
 }