+++ Source/EventFlow.ReadStores.MsSql/Attributes/MsSqlReadModelIdentityColumnAttribute.cs
@@ -21,10 +21,13 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
-namespace EventFlow.ReadStores.MsSql
+
+using System;
+
+namespace EventFlow.ReadStores.MsSql.Attributes
 {
-    public interface IMssqlReadModelStore<TReadModel> : IReadModelStore<TReadModel>
-        where TReadModel : class, IReadModel, new()
+    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field)]
+    public class MsSqlReadModelIdentityColumnAttribute : Attribute
     {
     }
 }