+++ Source/EventFlow.ReadStores.MsSql/MssqlReadModelStore.cs
@@ -21,37 +21,65 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Core;
+using EventFlow.Extensions;
 using EventFlow.Logs;
 using EventFlow.MsSql;
-using EventFlow.Queries;
+using EventFlow.ReadStores.MsSql.Attributes;
+
+#pragma warning disable 618
 
 namespace EventFlow.ReadStores.MsSql
 {
     public class MssqlReadModelStore<TReadModel> :
         ReadModelStore<TReadModel>,
         IMssqlReadModelStore<TReadModel>
-        where TReadModel : class, IMssqlReadModel, new()
+        where TReadModel : class, IReadModel, new()
     {
         private readonly IMsSqlConnection _connection;
-        private readonly IQueryProcessor _queryProcessor;
         private readonly IReadModelSqlGenerator _readModelSqlGenerator;
+        private static readonly Func<TReadModel, int?> GetVersion;
+        private static readonly Action<TReadModel, int?> SetVersion;
+
+        static MssqlReadModelStore()
+        {
+            var propertyInfos = typeof (TReadModel)
+                .GetProperties(BindingFlags.Instance | BindingFlags.Public);
+
+            var versionPropertyInfo = propertyInfos
+                .SingleOrDefault(p => p.GetCustomAttribute<MsSqlReadModelVersionColumnAttribute>() != null);
+            if (versionPropertyInfo == null)
+            {
+                versionPropertyInfo = propertyInfos.SingleOrDefault(p => p.Name == "LastAggregateSequenceNumber");
+            }
+
+            if (versionPropertyInfo == null)
+            {
+                GetVersion = rm => null as int?;
+                SetVersion = (rm, v) => { };
+            }
+            else
+            {
+                GetVersion = rm => (int?)versionPropertyInfo.GetValue(rm);
+                SetVersion = (rm, v) => versionPropertyInfo.SetValue(rm, v);
+            }
+        }
 
         public MssqlReadModelStore(
             ILog log,
             IMsSqlConnection connection,
-            IQueryProcessor queryProcessor,
             IReadModelSqlGenerator readModelSqlGenerator)
             : base(log)
         {
             _connection = connection;
-            _queryProcessor = queryProcessor;
             _readModelSqlGenerator = readModelSqlGenerator;
         }
 
@@ -61,32 +89,44 @@ namespace EventFlow.ReadStores.MsSql
             Func<IReadModelContext, IReadOnlyCollection<IDomainEvent>, ReadModelEnvelope<TReadModel>, CancellationToken, Task<ReadModelEnvelope<TReadModel>>> updateReadModel,
             CancellationToken cancellationToken)
         {
-            // TODO: Transaction
-
             foreach (var readModelUpdate in readModelUpdates)
             {
+                IMssqlReadModel mssqlReadModel;
+
                 var readModelNameLowerCased = typeof(TReadModel).Name.ToLowerInvariant();
                 var readModelEnvelope = await GetAsync(readModelUpdate.ReadModelId, cancellationToken).ConfigureAwait(false);
                 var readModel = readModelEnvelope.ReadModel;
                 var isNew = readModel == null;
+
                 if (readModel == null)
                 {
-                    readModel = new TReadModel
-                        {
-                            AggregateId = readModelUpdate.ReadModelId,
-                            CreateTime = readModelUpdate.DomainEvents.First().Timestamp,
-                        };
+                    readModel = new TReadModel();
+                    mssqlReadModel = readModel as IMssqlReadModel;
+                    if (mssqlReadModel != null)
+                    {
+                        mssqlReadModel.AggregateId = readModelUpdate.ReadModelId;
+                        mssqlReadModel.CreateTime = readModelUpdate.DomainEvents.First().Timestamp;
+                    }
+                    readModelEnvelope = ReadModelEnvelope<TReadModel>.With(readModel);
                 }
 
                 readModelEnvelope = await updateReadModel(
                     readModelContext,
                     readModelUpdate.DomainEvents,
-                    ReadModelEnvelope<TReadModel>.With(readModel, readModel.LastAggregateSequenceNumber),
+                    readModelEnvelope,
                     cancellationToken)
                     .ConfigureAwait(false);
 
-                readModel.UpdatedTime = DateTimeOffset.Now;
-                readModel.LastAggregateSequenceNumber = (int) readModelEnvelope.Version.GetValueOrDefault();
+                mssqlReadModel = readModel as IMssqlReadModel;
+                if (mssqlReadModel != null)
+                {
+                    mssqlReadModel.UpdatedTime = DateTimeOffset.Now;
+                    mssqlReadModel.LastAggregateSequenceNumber = (int)readModelEnvelope.Version.GetValueOrDefault();
+                }
+                else
+                {
+                    SetVersion(readModel, (int?) readModelEnvelope.Version);
+                }
 
                 var sql = isNew
                     ? _readModelSqlGenerator.CreateInsertSql<TReadModel>()
@@ -102,19 +142,31 @@ namespace EventFlow.ReadStores.MsSql
 
         public override async Task<ReadModelEnvelope<TReadModel>> GetAsync(string id, CancellationToken cancellationToken)
         {
-            var readModelNameLowerCased = typeof(TReadModel).Name.ToLowerInvariant();
+            var readModelType = typeof (TReadModel);
+            var readModelNameLowerCased = readModelType.Name.ToLowerInvariant();
             var selectSql = _readModelSqlGenerator.CreateSelectSql<TReadModel>();
             var readModels = await _connection.QueryAsync<TReadModel>(
                 Label.Named(string.Format("mssql-fetch-read-model-{0}", readModelNameLowerCased)),
                 cancellationToken,
                 selectSql,
-                new { AggregateId = id })
+                new { EventFlowReadModelId = id })
                 .ConfigureAwait(false);
+
             var readModel = readModels.SingleOrDefault();
 
-            return readModel == null
-                ? ReadModelEnvelope<TReadModel>.Empty
-                : ReadModelEnvelope<TReadModel>.With(readModel, readModel.LastAggregateSequenceNumber);
+            if (readModel == null)
+            {
+                Log.Verbose(() => $"Could not find any MSSQL read model '{readModelType.PrettyPrint()}' with ID '{id}'");
+                return ReadModelEnvelope<TReadModel>.Empty;
+            }
+
+            var readModelVersion = GetVersion(readModel);
+
+            Log.Verbose(() => $"Foud MSSQL read model '{readModelType.PrettyPrint()}' with ID '{readModelVersion}'");
+
+            return readModelVersion.HasValue
+                ? ReadModelEnvelope<TReadModel>.With(readModel, readModelVersion.Value)
+                : ReadModelEnvelope<TReadModel>.With(readModel);
         }
 
         public override async Task DeleteAllAsync(CancellationToken cancellationToken)