+++ Source/EventFlow.Tests/IntegrationTests/ReadStores/QueryHandlers/InMemoryThingyGetMessagesQueryHandler.cs
@@ -21,10 +21,35 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
-namespace EventFlow.ReadStores.MsSql
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using EventFlow.Queries;
+using EventFlow.ReadStores.InMemory;
+using EventFlow.TestHelpers.Aggregates.Entities;
+using EventFlow.TestHelpers.Aggregates.Queries;
+using EventFlow.Tests.IntegrationTests.ReadStores.ReadModels;
+
+namespace EventFlow.Tests.IntegrationTests.ReadStores.QueryHandlers
 {
-    public interface IMssqlReadModelStore<TReadModel> : IReadModelStore<TReadModel>
-        where TReadModel : class, IReadModel, new()
+    public class InMemoryThingyGetMessagesQueryHandler : IQueryHandler<ThingyGetMessagesQuery, IReadOnlyCollection<ThingyMessage>>
     {
+        private readonly IInMemoryReadStore<InMemoryThingyMessageReadModel> _readStore;
+
+        public InMemoryThingyGetMessagesQueryHandler(
+            IInMemoryReadStore<InMemoryThingyMessageReadModel> readStore)
+        {
+            _readStore = readStore;
+        }
+
+        public async Task<IReadOnlyCollection<ThingyMessage>> ExecuteQueryAsync(ThingyGetMessagesQuery query, CancellationToken cancellationToken)
+        {
+            var readModels = await _readStore.FindAsync(rm => rm.ThingyId == query.ThingyId, cancellationToken).ConfigureAwait(false);
+            return readModels
+                .Select(rm => rm.ToThingyMessage())
+                .ToList();
+        }
     }
 }