+++ Source/EventFlow.Tests/IntegrationTests/ReadStores/ReadModels/InMemoryThingyMessageReadModel.cs
@@ -22,30 +22,35 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
 
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using EventFlow.Queries;
-using EventFlow.ReadStores.InMemory;
+using EventFlow.Aggregates;
+using EventFlow.ReadStores;
 using EventFlow.TestHelpers.Aggregates;
-using EventFlow.TestHelpers.Aggregates.Queries;
+using EventFlow.TestHelpers.Aggregates.Entities;
+using EventFlow.TestHelpers.Aggregates.Events;
 
-namespace EventFlow.Tests.IntegrationTests.ReadStores
+namespace EventFlow.Tests.IntegrationTests.ReadStores.ReadModels
 {
-    public class InMemoryThingyGetQueryHandler : IQueryHandler<ThingyGetQuery, Thingy>
+    public class InMemoryThingyMessageReadModel : IReadModel,
+        IAmReadModelFor<ThingyAggregate, ThingyId, ThingyMessageAddedEvent>
     {
-        private readonly IInMemoryReadStore<InMemoryThingyReadModel> _readStore;
+        public ThingyId ThingyId { get; private set; }
+        public ThingyMessageId ThingyMessageId { get; private set; }
+        public string Message { get; private set; }
 
-        public InMemoryThingyGetQueryHandler(
-            IInMemoryReadStore<InMemoryThingyReadModel> readStore)
+        public void Apply(IReadModelContext context, IDomainEvent<ThingyAggregate, ThingyId, ThingyMessageAddedEvent> domainEvent)
         {
-            _readStore = readStore;
+            ThingyId = domainEvent.AggregateIdentity;
+
+            var thingyMessage = domainEvent.AggregateEvent.ThingyMessage;
+            ThingyMessageId = thingyMessage.Id;
+            Message = thingyMessage.Message;
         }
 
-        public async Task<Thingy> ExecuteQueryAsync(ThingyGetQuery query, CancellationToken cancellationToken)
+        public ThingyMessage ToThingyMessage()
         {
-            var readModels = await _readStore.FindAsync(rm => rm.ThingyId == query.ThingyId, cancellationToken).ConfigureAwait(false);
-            return readModels.SingleOrDefault()?.ToThingy();
+            return new ThingyMessage(
+                ThingyMessageId,
+                Message);
         }
     }
 }